apply code review suggestions
