+++ bitcoin/client.go
@@ -1,0 +1,12 @@
+package bitcoin
+
+import (
+	"github.com/btcsuite/btcd/chaincfg/chainhash"
+	"github.com/btcsuite/btcd/wire"
+)
+
+// BitcoinClient defines the interface with only the functions we need.
+type BitcoinClient interface {
+	SendRawTransaction(tx *wire.MsgTx, allowHighFees bool) (*chainhash.Hash, error)
+	Shutdown()
+}
